+++ tests/detectors/test_detectors.py
@@ -26,13 +26,11 @@ from pyxel.detectors import (
     MKIDGeometry,
 )
 
-try:
-    # Check if library 'asdf' is installed
-    import asdf  # noqa: F401
-except ImportError:
-    WITH_ASDF = False
-else:
-    WITH_ASDF = True
+# This is equivalent to 'import asdf'
+asdf = pytest.importorskip(
+    "asdf",
+    reason="Package 'asdf' is not installed. Use 'pip install asdf'",
+)
 
 
 @pytest.fixture(
@@ -184,7 +182,6 @@ def test_to_from_hdf5(detector: Union[CCD, CMOS, MKID, APD], tmp_path: Path):
     assert new_detector == detector
 
 
-@pytest.mark.skipif(WITH_ASDF is False, reason="Missing library 'asdf'")
 def test_to_from_asdf(detector: Union[CCD, CMOS, MKID, APD], tmp_path: Path):
     """Test methods `Detector.to_asdf' and `Detector.from_asdf`."""
     filename = tmp_path / f"{detector.__class__.__name__}.asdf"
@@ -217,9 +214,6 @@ def test_to_from_filename(
 ):
     """Test methods `Detector.load` and `Detector.save`."""
     full_filename: Path = tmp_path / filename
-
-    if full_filename.suffix == ".asdf" and not WITH_ASDF:
-        pytest.skip(reason="Missing library 'asdf'")
 
     # Save
     detector.save(full_filename)