+++ CHANGELOG.md
@@ -18,7 +18,8 @@ Pyxel doesn't use SemVer anymore, since most minor releases have at least minor 
 ### Models
 
 ### Others
-
+* Improved tests when an optional dependency is not installed.
+  (See [!866](https://gitlab.com/esa/pyxel/-/merge_requests/866)).
 
 ## 2.0 / 2024-02-22